# frontend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/snapshots/position/{positionId}": {
      "get": {
        "operationId": "PositionSnapshotController_getPositionSnapshots",
        "parameters": [
          {
            "name": "positionId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "5ngdutp5g8cTbGFD4dV2hgGQPkPonp4gp32ZxPaMuiP",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2025-01-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get snapshots for position",
        "tags": [
          "Position Snapshots"
        ]
      }
    },
    "/snapshots/recent": {
      "get": {
        "operationId": "PositionSnapshotController_getRecentSnapshots",
        "parameters": [
          {
            "name": "days",
            "required": false,
            "in": "query",
            "schema": {
              "example": 7,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get recent snapshots for all positions",
        "tags": [
          "Position Snapshots"
        ]
      }
    },
    "/snapshots/statistics": {
      "get": {
        "operationId": "PositionSnapshotController_getPeriodStatistics",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2025-01-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get statistics for period",
        "tags": [
          "Position Snapshots"
        ]
      }
    },
    "/snapshots/trigger": {
      "post": {
        "operationId": "PositionSnapshotController_triggerSnapshot",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Manually trigger snapshot creation",
        "tags": [
          "Position Snapshots"
        ]
      }
    },
...
  },
  "info": {
    "title": "Liquidity Bot API",
    "description": "API для управления ликвидностью на Raydium CLMM",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
...
  }
}